feat: add book export and import

+++ BookManagerApp/Models/BookManager.cs
@@ -1,12 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Diagnostics;
+using System.Globalization;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using System.Windows.Shapes;
 using BookManagerApp.DataAccessLayer;
 using BookManagerApp.Views;
+using CsvHelper.Configuration;
+using CsvHelper;
 using Microsoft.EntityFrameworkCore;
 using static System.Reflection.Metadata.BlobBuilder;
 
@@ -14,11 +20,14 @@ namespace BookManagerApp.Models
 {
     public class BookManager
     {
-        public async static Task<bool> AddNewBook(Book book)
+        static int counter = 0;
+        public async static Task<bool> AddNewBook(Book book, User user)
         {
             using var context = new AppDbContext();
             var temp = await context.Books
-                .FirstOrDefaultAsync(b => b.Title == book.Title);
+                .FirstOrDefaultAsync(b => b.Title == book.Title 
+                                          && b.Username == user.Username
+                                          && b.Author == book.Author);
             if (temp != null)
             {
                 return false;
@@ -31,7 +40,13 @@ namespace BookManagerApp.Models
         public async static Task DeleteBook(Book book)
         {
             using var context = new AppDbContext();
-            context.Books.Remove(book);
+            var existing = await context.Books
+                .FirstOrDefaultAsync(b => b.Id == book.Id);
+            if (existing != null)
+            {
+                context.Entry(existing).State = EntityState.Detached;
+                context.Books.Remove(book);
+            }
             await context.SaveChangesAsync();
         }
 
@@ -87,5 +102,95 @@ namespace BookManagerApp.Models
                             && (b.Genre == null ? false : b.Genre.Contains(genreFilter)));
             return newBooks;
         }
+
+        public async static Task ImportBooksFromCSV(string filePath, User user)
+        {
+            using var context = new AppDbContext();
+            using (var sourceStream =
+            new FileStream(
+                filePath,
+                FileMode.Open, FileAccess.Read, FileShare.Read,
+                bufferSize: 4096, useAsync: true))
+            {
+                using (var reader = new StreamReader(sourceStream))
+                {
+                    using (var csv = new CsvReader(reader, new CsvConfiguration(CultureInfo.InvariantCulture)
+                    {
+                        Delimiter = ",",
+                        Quote = '"'
+                    }))
+                    {
+                        await foreach (var record in csv.GetRecordsAsync<dynamic>())
+                        {
+                            var dict = (IDictionary<string, object>)record;
+
+                            int rating;
+                            Int32.TryParse(dict["My Rating"]?.ToString(), out rating);
+
+                            DateTime dateTime;
+                            DateTime.TryParse(dict["Date Read"]?.ToString(), out dateTime);
+
+                            if (await context.Books
+                                .FirstOrDefaultAsync(b => b.Title == dict["Title"].ToString() 
+                                                          && b.Username == user.Username
+                                                          && b.Author == dict["Author"].ToString()) != null) 
+                            {
+                                continue;
+                            }
+                            var book = new Book(dict["Author"].ToString() ?? $"Author{counter}",
+                                                    dict["Title"].ToString() ?? $"Book{counter++}",
+                                                    convertGoodReadsBookshelf(dict["Exclusive Shelf"]?.ToString() ?? "to-read"),
+                                                    user.Username,
+                                                    dict["My Rating"].ToString() == "" ? null : rating,
+                                                    dict["Date Read"].ToString() == "" ? null : dateTime,
+                                                    "Unknown");
+                            user.Books.Add(book);
+                            context.Add(book);
+                        }
+                    }
+                }
+                await context.SaveChangesAsync();
+            }
+        }
+
+        public static async Task ExportBooksToCSV(string filePath, User user)
+        {
+            using var context = new AppDbContext();
+            using (var sourceStream =
+            new FileStream(
+                filePath,
+                FileMode.Create, FileAccess.Write, FileShare.None,
+                bufferSize: 4096, useAsync: true))
+            {
+                using (var writer = new StreamWriter(sourceStream))
+                {
+                    using (var csv = new CsvWriter(writer, new CsvConfiguration(CultureInfo.InvariantCulture)
+                    {
+                        Delimiter = ",",
+                        Quote = '"'
+                    }))
+                    {
+                        var infoToExport = user.Books
+                            .Select(b => new {
+                                b.Title,
+                                b.Author,
+                                b.MyRating,
+                                b.FinishDate,
+                                b.Bookshelf,
+                                b.Genre
+                            });
+                        await csv.WriteRecordsAsync(infoToExport);
+                    }
+                }
+            }
+        }
+
+        private static Func<string, string> convertGoodReadsBookshelf = input => input switch
+        {
+            "read" => "Completed",
+            "currently-reading" => "Currently Reading",
+            "to-read" => "Want To Read",
+            _ => "Want To Read"
+        };
     }
 }
